requirements.txt
Django==5.1.2
django-ninja==1.3.0
scipy==1.13.1
spiceypy==6.0.0
numpy==1.26.4
pymap3d==3.1.0
arrow==1.3.0
gdal==3.10